Move more functions to utils.go
